+++ internal/config/parsed_rule.go
@@ -116,7 +116,7 @@ func defaultMatchStates(cmd ContextCommandVal) []string {
 }
 
 func baseRules(staticRules []staticRule, proms []*promapi.FailoverGroup, match []Match) (rules []*parsedRule) {
-	const checksPerProm = 10 // Number of checks registered per Prometheus server below.
+	const checksPerProm = 11 // Number of checks registered per Prometheus server below.
 	rules = make([]*parsedRule, 0, len(staticRules)+(len(proms)*checksPerProm))
 	for _, sr := range staticRules {
 		rules = append(rules, baseParsedRule(match, sr.name, sr.checker, nil))
@@ -128,6 +128,7 @@ func baseRules(staticRules []staticRule, proms []*promapi.FailoverGroup, match [
 			baseParsedRule(match, checks.RateCheckName, checks.NewRateCheck(p), p.Tags()),
 			baseParsedRule(match, checks.SeriesCheckName, checks.NewSeriesCheck(p), p.Tags()),
 			baseParsedRule(match, checks.VectorMatchingCheckName, checks.NewVectorMatchingCheck(p), p.Tags()),
+			baseParsedRule(match, checks.OffsetCheckName, checks.NewOffsetCheck(p), p.Tags()),
 			baseParsedRule(match, checks.RangeQueryCheckName, checks.NewRangeQueryCheck(p, 0, "", checks.Warning), p.Tags()),
 			baseParsedRule(match, checks.RuleDuplicateCheckName, checks.NewRuleDuplicateCheck(p), p.Tags()),
 			baseParsedRule(match, checks.LabelsConflictCheckName, checks.NewLabelsConflictCheck(p), p.Tags()),